# Conflicts:
#	UJCSystem/ITJakub.BatchImport.Client/App.config
#	UJCSystem/ITJakub.BatchImport.Client/ITJakub.BatchImport.Client.csproj
#	UJCSystem/ITJakub.BatchImport.Client/packages.config
#	UJCSystem/ITJakub.Lemmatization.Service.Test/app.config
#	UJCSystem/ITJakub.Lemmatization.Service/Web.config
#	UJCSystem/ITJakub.Web.Hub/Localization/BibliographyModule/BibliographyModule.cs-CZ.json
#	UJCSystem/ITJakub.Web.Hub/Localization/BibliographyModule/BibliographyModule.en.json
#	UJCSystem/ITJakub.Web.Hub/wwwroot/content/BibliographyConfiguration/configuration.json
#	UJCSystem/UJCSystem.sln
#	UJCSystem/Vokabular.DataEntities/DataEntitiesContainerRegistration.cs
#	UJCSystem/Vokabular.DataEntities/Database/Entities/Project.hbm.xml
#	UJCSystem/Vokabular.MainService.Core/MainServiceCoreContainerRegistration.cs
#	UJCSystem/Vokabular.MainService.Core/Managers/PermissionManager.cs
#	UJCSystem/Vokabular.MainService.Core/Managers/RoleManager.cs
#	UJCSystem/Vokabular.MainService.Core/Managers/UserManager.cs
#	UJCSystem/Vokabular.MainService.Core/Works/Users/UpdateCurrentUserWork.cs
#	UJCSystem/Vokabular.MainService.Core/Works/Users/UpdateUserPasswordWork.cs
#	UJCSystem/Vokabular.MainService.DataContracts/Clients/MainServiceRestClient.cs
#	UJCSystem/Vokabular.MainService.DataContracts/Contracts/Search/SearchResultContract.cs
#	UJCSystem/Vokabular.MainService/Startup.cs
#	UJCSystem/Vokabular.MainService/appsettings.LocalDebug.json
#	UJCSystem/Vokabular.MainService/appsettings.json

+++ UJCSystem/Vokabular.MainService.Core/Works/Permission/DeleteRoleWork.cs
@@ -1,4 +1,4 @@
-﻿using Vokabular.Authentication.DataContracts;
+﻿using Ridics.Authentication.DataContracts;
 using Vokabular.DataEntities.Database.Repositories;
 using Vokabular.MainService.Core.Communication;
 using Vokabular.Shared.DataEntities.UnitOfWork;
@@ -21,9 +21,12 @@ namespace Vokabular.MainService.Core.Works.Permission
         protected override void ExecuteWorkImplementation()
         {
             var group = m_permissionRepository.FindGroupByExternalId(m_roleId);
-            m_permissionRepository.Delete(group);
-            m_permissionRepository.Flush();
-
+            if (group != null)
+            {
+                m_permissionRepository.Delete(group);
+                m_permissionRepository.Flush();
+            }
+            
             var client = m_communicationProvider.GetAuthRoleApiClient();
             client.HttpClient.DeleteItemAsync<RoleContract>(m_roleId).GetAwaiter().GetResult();
         }
